cypress: BLN: Use wakelock for breathing when timeout is set
It will enable wakelock automatically when breathing is enabled
with notification-timeout set and fading breathing configuration.
This way the breathing-effect doesn't freeze by preventing the
device to go to deep sleep until it times out. It won't use wakelock
unneccesary when no fading and single line configuration is used.

The use of wakelock can be overriden by writing 0 (disable) or 1
(enable) to the following path:
/sys/devices/virtual/misc/backlightnotification/breathing_use_wakelock

Change-Id: I7fb352e505e2adc66ab709ab3a740dd89071f567